requirements.txt
requests
beautifulsoup4
curl_cffi
pytz
python-dotenv